% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SsHDFS.R
\name{hdfs.cat}
\alias{hdfs.cat}
\title{send the content of a remote hdfs file via a pipe}
\usage{
hdfs.cat(src.path, host = "analytix", user = "")
}
\arguments{
\item{src.path}{HDFS location of the source file}

\item{host}{remote host}

\item{user}{remote user}
}
\value{
an open pipe to the data.
}
\description{
send the content of a remote hdfs file via a pipe
}
\examples{
read.csv(hdfs.cat("input.txt"))
}

