% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SsHDFS.R
\name{ssh.pipe}
\alias{ssh.pipe}
\title{Execute a remote command via a ssh pipe}
\usage{
ssh.pipe(cmd, host = "localhost", user = "", open.mode = "")
}
\arguments{
\item{cmd}{command string to be executed on the remote shell}

\item{host}{hostname (or ssh config shortcut) to reach the remote execution host. defaults to localhost}

\item{user}{username on remote system}

\item{open.mode}{open mode for the pipe. defaults to text mode but can be changed with "rb" to read binary}
}
\value{
an open pipe to the remote process.
}
\description{
Execute a remote command via a ssh pipe
}
\examples{
readLines(ssh.pipe("ls", "analytix"))
}

