% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SsHDFS.R
\name{hdfs.copyToLocal}
\alias{hdfs.copyToLocal}
\title{Copy a file from the remote HDSF area to the current directory}
\usage{
hdfs.copyToLocal(src.path, host = "analytix", user = "",
  dest.path = src.path)
}
\arguments{
\item{src.path}{source pathname in HDFS}

\item{host}{remote execution host}

\item{user}{remote execution user}

\item{dest.path}{local filename (defaults to remote file name)}
}
\value{
nothing
}
\description{
Copy a file from the remote HDSF area to the current directory
}
\examples{
hdfs.copyToLocal("results.txt")
}

