% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SsHDFS.R
\name{hdfs.du}
\alias{hdfs.du}
\title{Report remote HDSF disk usage}
\usage{
hdfs.du(args, host = "analytix", user = "")
}
\arguments{
\item{args}{argument string passed to remote hdfs du command}

\item{host}{remote execution host}

\item{user}{remote execution user}
}
\value{
return data frame with the raw and user disk volume per queried entity
}
\description{
Report remote HDSF disk usage
}
\examples{
hdfs.du("/project/awg")
}

