% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SsHDFS.R
\name{hdfs.ls}
\alias{hdfs.ls}
\title{Scan a directory and return relevant meta data as data frame}
\usage{
hdfs.ls(args, host = "analytix", user = "")
}
\arguments{
\item{args}{arguments passed to hdfs dfs commandline}

\item{host}{remote execution host}

\item{user}{remote execution user}
}
\value{
a data frame with the resulting file meta data per row
}
\description{
Scan a directory and return relevant meta data as data frame
}
\examples{
itmon <- hdfs.ls("-R /project/itmon")
}

